@@ -1,8 +1,14 @@
 module github.com/alekseiapa/tg-client
 
-go 1.16
+go 1.19
 
 require (
 	github.com/jarcoal/httpmock v1.0.8
 	github.com/stretchr/testify v1.7.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
 )
